ref: del stupid util file
